@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.1"
+    compileSdkVersion 33
+    buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "cafe.adriel.androidaudioconverter.sample"
-        minSdkVersion 16
-        targetSdkVersion 24
+        minSdkVersion 26
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
     }
@@ -19,6 +19,8 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:24.2.1'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.core:core:1.7.0'
+
     compile project(':lib')
 }
